# ExpressApolloAPI/server/config/swagger.yaml
openapi: 3.0.4
info:
  title: Case Management API
  description: API สำหรับจัดการเคส เช่น เพิ่ม, อัปเดตสถานะ, ผลลัพธ์, เพิ่มโน้ต และดูประวัติ
  version: 1.0.0

servers:
  - url: http://172.16.0.12:4000
    description: Local development server

security:
  - UserEmailHeader: []

tags:
  - name: User Accounts
  - name: Incidents
  - name: History

components:
  securitySchemes:
    UserEmailHeader:
      type: apiKey
      in: header
      name: user_email

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        user_email:
          type: string

    UserStatus:
      type: object
      properties:
        id:
          type: string
        user_email:
          type: string
        account_status:
          type: string

    Incident:
      type: object
      properties:
        id:
          type: string
        alert_id:
          type: string
        alert_name:
          type: string
        alert_status:
          type: string
        case_result:
          type: string
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"

    IncidentList:
      type: array
      items:
        $ref: "#/components/schemas/Incident"

    Note:
      type: object
      properties:
        id:
          type: string
        action:
          type: string
        content:
          type: string
        created_at:
          type: string
          format: date-time

    HistoryEntry:
      type: object
      properties:
        authentication:
          type: object
          properties:
            user_email:
              type: string
            name:
              type: string
            id:
              type: string
        user_agent:
          type: string
        ip_address:
          type: string
        action:
          type: string
        case:
          oneOf:
            - $ref: "#/components/schemas/Incident"
            - $ref: "#/components/schemas/Note"
          description: ข้อมูลที่เกี่ยวข้องกับ action เช่น Incident หรือ Note

    Error:
      type: object
      properties:
        error:
          type: string

paths:
  # /accounts/users:
  #   get:
  #     summary: Get list of user accounts
  #     tags: [User Accounts]
  #     responses:
  #       "200":
  #         description: รายชื่อผู้ใช้งานทั้งหมด
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/IncidentList"
  #       "500":
  #         description: เกิดข้อผิดพลาดจากฝั่งเซิร์ฟเวอร์
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/Error"


  /accounts/unlock:
    put:
      summary: Unlock user account(s) - รองรับส่งทีละตัวหรือหลายตัว
      tags: [User Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [user_email]
                  properties:
                    user_email:
                      type: string
                      example: "abc@gmail.com"
                - type: object
                  required: [users]
                  properties:
                    users:
                      type: array
                      items:
                        type: object
                        required: [user_email]
                        properties:
                          user_email:
                            type: string
                      example:
                        - user_email: "abc@gmail.com"
                        - user_email: "def@gmail.com"
      responses:
        "200":
          description: Account(s) unlocked successfully with notes added
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      oneOf:
                        - type: object
                          properties:
                            id:
                              type: string
                            user_email:
                              type: string
                            account_status:
                              type: string
                            note:
                              $ref: "#/components/schemas/Note"
                        - type: object
                          properties:
                            user_email:
                              type: string
                            error:
                              type: string
                example:
                  results:
                    - id: "user-001"
                      user_email: "abc@gmail.com"
                      account_status: "Unlocked"
                      note:
                        id: "note-789"
                        action: "Unlocked"
                        content: "Account unlocked by admin"
                        created_at: "2025-06-22T08:00:00Z"
                    - user_email: "notfound@example.com"
                      error: "User not found with given email"
        "400":
          description: Missing or invalid user_email(s)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Missing or invalid 'user_email'"
        "404":
          description: User not found or unlock failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "User not found with given email"
        "500":
          description: Internal server error or failed to unlock one or more accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Failed to unlock one or more accounts"

  # /incidents:
  #   get:
  #     summary: ดึงรายชื่อ incidents ทั้งหมด
  #     tags: [Incidents]
  #     responses:
  #       "200":
  #         description: รายชื่อ incidents
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/IncidentList"
  #       "500":
  #         description: เกิดข้อผิดพลาดในการดึงข้อมูล
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/Error"

  # /incident:
  #   put:
  #     summary: ดึง incident ตาม ID (ผ่าน request body)
  #     tags: [Incidents]
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             required: [id]
  #             properties:
  #               id:
  #                 type: string
  #                 example: "incident"
  #     responses:
  #       "200":
  #         description: ข้อมูล incident ตาม ID
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/Incident"
  #       "400":
  #         description: id ไม่ถูกต้อง
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/Error"
  #       "500":
  #         description: เกิดข้อผิดพลาดในการดึงข้อมูล
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/Error"

  /closedAlertStatus:
    put:
      summary: ปิด alert_status ของ incident พร้อมเพิ่ม note อัตโนมัติ
      tags: [Incidents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [id, alert_status]
                  properties:
                    id:
                      type: string
                    alert_status:
                      type: string
                      enum: [Closed]
                - type: object
                  required: [incidents]
                  properties:
                    incidents:
                      type: array
                      items:
                        type: object
                        required: [id, alert_status]
                        properties:
                          id:
                            type: string
                          alert_status:
                            type: string
                            enum: [Closed]
            examples:
              SingleIncident:
                summary: ปิดเคสเดียว
                value:
                  id: "incident-001"
                  alert_status: "Closed"
              MultipleIncidents:
                summary: ปิดหลายเคส
                value:
                  incidents:
                    - id: "incident01"
                      alert_status: "Closed"
                    - id: "incident02"
                      alert_status: "Closed"
      responses:
        "200":
          description: อัปเดต alert_status สำเร็จและเพิ่ม note
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        updated:
                          type: boolean
                        alert_status:
                          type: string
                        note:
                          $ref: "#/components/schemas/Note"
                        error:
                          type: string
              example:
                results:
                  - id: "incident-001"
                    updated: true
                    alert_status: "Closed"
                    note:
                      id: "note-999"
                      action: "Closed"
                      content: "Incident was Closed"
                      created_at: "2025-06-22T08:00:00Z"
        "400":
          description: ข้อมูลไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: เกิดข้อผิดพลาด
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /updateCaseResult:
    put:
      summary: อัปเดต case_result ของ incident พร้อมเพิ่ม note
      tags: [Incidents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [id, case_result, reason]
                  properties:
                    id:
                      type: string
                    case_result:
                      type: string
                      enum: [TruePositive, FalsePositive, Benign, Information]
                    reason:
                      type: string
                - type: object
                  required: [incidents]
                  properties:
                    incidents:
                      type: array
                      items:
                        type: object
                        required: [id, case_result, reason]
                        properties:
                          id:
                            type: string
                          case_result:
                            type: string
                            enum:
                              [TruePositive, FalsePositive, Benign, Information]
                          reason:
                            type: string
            examples:
              SingleCaseUpdate:
                summary: อัปเดตผลเคสเดียว
                value:
                  id: "incident-003"
                  case_result: "TruePositive"
                  reason: "Verified as true positive"
              MultipleCaseUpdate:
                summary: อัปเดตหลายเคส
                value:
                  incidents:
                    - id: "incident-003"
                      case_result: "TruePositive"
                      reason: "Verified as threat"
                    - id: "incident-004"
                      case_result: "Benign"
                      reason: "Legitimate admin activity"
      responses:
        "200":
          description: อัปเดต case_result สำเร็จและเพิ่ม note
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        updated:
                          type: boolean
                        case_result:
                          type: string
                        note:
                          $ref: "#/components/schemas/Note"
                        error:
                          type: string
              example:
                results:
                  - id: "incident-003"
                    updated: true
                    case_result: "TruePositive"
                    note:
                      id: "note-998"
                      action: "Re-Investigated"
                      content: "Verified as threat"
                      created_at: "2025-06-22T08:10:00Z"
        "400":
          description: ข้อมูลไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: เกิดข้อผิดพลาด
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /history:
    get:
      summary: ดึงประวัติการเปลี่ยนแปลงทั้งหมด
      tags: [History]
      responses:
        "200":
          description: รายการประวัติ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryEntry"
        "500":
          description: เกิดข้อผิดพลาดในการอ่านประวัติ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"